# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Stray dog reports
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all stray dog reports
      responses:
        "200":
          description: List of reports.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new stray dog report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportInput"
      responses:
        "201":
          description: Report created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single stray dog report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a stray dog report status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportInput"
      responses:
        "200":
          description: Updated report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a stray dog report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted.
        "404":
          description: Report not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /reports/stats/summary:
    get:
      operationId: getReportsSummary
      tags: [reports]
      summary: Aggregated summary of all stray dog reports
      responses:
        "200":
          description: Stats summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsSummary"

  /reports/recent:
    get:
      operationId: getRecentReports
      tags: [reports]
      summary: Get the most recent stray dog reports
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        "200":
          description: Recent reports.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Report:
      type: object
      required:
        - id
        - location
        - description
        - status
        - createdAt
      properties:
        id:
          type: integer
        location:
          type: string
        description:
          type: string
        imagePath:
          type: ["string", "null"]
          description: Object path of the uploaded image, or null if no image was provided.
        status:
          type: string
          enum: [reported, in_progress, rescued]
        createdAt:
          type: string
          format: date-time

    CreateReportInput:
      type: object
      required:
        - location
        - description
      properties:
        location:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        imagePath:
          type: ["string", "null"]

    UpdateReportInput:
      type: object
      properties:
        status:
          type: string
          enum: [reported, in_progress, rescued]

    ReportsSummary:
      type: object
      required:
        - total
        - reported
        - inProgress
        - rescued
        - last7Days
      properties:
        total:
          type: integer
        reported:
          type: integer
        inProgress:
          type: integer
        rescued:
          type: integer
        last7Days:
          type: integer

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string