# lib/api-zod/src/generated/types/createReportInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReportInput {
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  imagePath?: string | null;
}