# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from "./reportStatus";

export interface Report {
  id: number;
  location: string;
  description: string;
  /**
   * Object path of the uploaded image, or null if no image was provided.
   * @nullable
   */
  imagePath?: string | null;
  status: ReportStatus;
  createdAt: Date;
}