# lib/api-zod/src/generated/types/reportsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportsSummary {
  total: number;
  reported: number;
  inProgress: number;
  rescued: number;
  last7Days: number;
}